requirements.txt
apipkg==1.5
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4
boto3==1.14.29
botocore==1.17.29
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
cryptography==3.0
docutils==0.15.2
execnet==1.7.1
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
jmespath==0.10.0
keyring==21.2.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
Naked==0.1.31
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.9.8
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.3.5
pytest-dependency==0.5.1
pytest-forked==1.2.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-parallel==0.1.0
pytest-xdist==1.33.0
python-dateutil==2.8.1
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
selenium==3.141.0
selenium-wire==1.2.3
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
tblib==1.6.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9



